[refactor] #30 ListItem.kt 이동
